Django shop | Session 9 | Send sms code and created a method for expiration code

+++ Django 3/A/accounts/views.py
@@ -2,24 +2,25 @@ from django.shortcuts import render, redirect
 from django.views import View
 from .forms import UserRegistrationForm, VerifyCodeForm
 import random
-from utils import send_top_code
+from utils import send_otp_code
 from .models import OtpCode, User
 from django.contrib import messages
 
 
 class UserRegisterView(View):
     form_class = UserRegistrationForm
+    template_name = 'accounts/register.html'
 
     def get(self, request):
         form = self.form_class
-        return render(request, 'accounts/register.html', {'form': form})
+        return render(request, self.template_name, {'form': form})
 
     def post(self, request):
         form = self.form_class(request.POST)
         if form.is_valid():
             cd = form.cleaned_data
             random_code = random.randint(1000, 9999)
-            send_top_code(cd['phone'], random_code)
+            send_otp_code(cd['phone'], random_code)
             OtpCode.objects.create(phone_number=cd['phone'], code=random_code)
             request.session['user_registration_info'] = {
                 'phone_number': cd['phone'],
@@ -29,7 +30,7 @@ class UserRegisterView(View):
             }
             messages.success(request, 'We sent you a code', 'success')
             return redirect('accounts:verify_code')
-        return redirect('home:home')
+        return render(request, self.template_name, {'form': form})
 
 
 class UserRegisterVerifyCodeView(View):
@@ -46,13 +47,16 @@ class UserRegisterVerifyCodeView(View):
         if form.is_valid():
             cd = form.cleaned_data
             if cd['code'] == code_instance.code:
-                User.objects.create_user(user_session['phone_number'], user_session['email'],
-                                         user_session['full_name'], user_session['password'])
-                code_instance.delete()
-                messages.success(request, 'You registered successfully.', 'success')
-                return redirect('home:home')
+                if code_instance.is_not_expired():
+                    User.objects.create_user(user_session['phone_number'], user_session['email'],
+                                             user_session['full_name'], user_session['password'])
+                    code_instance.delete()
+                    messages.success(request, 'You registered successfully.', 'success')
+                    return redirect('home:home')
+                else:
+                    messages.error(request, 'The code has expired. Please try again.', 'danger')
+                    return redirect('accounts:verify_code')
             else:
                 messages.error(request, 'This code is WRONG', 'danger')
                 return redirect('accounts:verify_code')
         return redirect('home:home')
-
